move all server details to the same file
